add alloc_grow, realloc with backoff
alloc_grow receives a min..max range of how much a memory region should be grown and tries a
backoff strategy when the max size fails, halving the additional memory until it finally fails
when it can't grow to the minimum number of elements.

This is named *_grow in anticipation of a *_shrink eventually which may needs a different strategy.

Use alloc_grow in trail, choice, frame and slot allocation.