Fault injection driver shell script
./tests/faultinject.sh iterates over all possible faults and restarts a test. When that test
crashed (by segv or abort) some logfiles are (stderr, stdout, backtrace) are kept.

The simple iteration is a quite inefficient test I have a better scheme in nobug, but that
needs system specific machinery (glibc access to backtraces).

For now this works, as by default it stops on the first crash.

There are the options:
  -k for keep going instead stopping
  -r for testing in reverse

Example:
 ./tests/faultinject.sh -k -r ./tpl -g "hanoiq(10),halt" samples/hanoi.pl