cdebug: print message one fault inject, include errno
Further code will rely on errno != 0 for handling errors. This allows more condensed error
handling, one can do a bunch of operations and only check for error at the end instead
checking for the success of every single operation.

This is not strictly C99 but POSIX but I am not aware of any system that does NOT do
this. Even if, then it will be easier to wrap the malloc/calloc calls or set errno at the
allocating function than handling all possible errors explicitly.